requirements.txt
aiogram==2.23.1
pymongo==4.5.0
python-dotenv~=1.0.1